daemon: support running without PolicyKit
Make the dependency on PolicyKit optional. In that case, don't
allow all methods and property access and fall back on the
standard D-Bus policies. NB: These are designed to work with
PolicyKit so most likely are not suitable at all. This is
designed for a special use case not intended for general use.

Since PolicyKit is the normal mode of operation, this feature
requires explicit opt-in, via the new `require-polkit` configure
option.